Create choose action file
